Adding a link to the info table in TOC/Grid view, with copy to clipboard and snackbar fetched async
This monkey-patches Aras' uiPopulateInfoTableWithItem to load HTML/CSS and add a
clickable link which copies the current item's URL to clipboard.

It also adds a 'bitAddictCopyToClipBoard' method to the Aras object which can
be used for any purpose.

The HTML/CSS is fetched asynchronously to avoid inlining it in javascript,
plus it shows how additional/customer HTML/CSS can be added to Aras.